Refactoring files
