+++ webapp/src/main/webapp/src/app/admin/sandbox/modal/reset-data.modal.html
