+++ webapp/src/main/webapp/src/app/admin/sandbox/model/sandbox-configuration-property.ts
