+++ webapp/src/main/webapp/src/app/admin/sandbox/pages/sandbox.component.html
@@ -6,7 +6,7 @@
       style="margin-top: -8px;"
       routerLink="/sandboxes/new"
     >
-      Create New Sandbox
+      {{ 'sandbox-config.title' | translate }}
     </a>
   </ng-template>
 </page-heading>
