+++ webapp/src/main/webapp/src/app/admin/sandbox/sandbox.module.ts
@@ -1,19 +1,19 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { SandboxConfigurationComponent } from './sandbox.component';
-import { SandboxConfigurationDetailsComponent } from './sandbox-details.component';
+import { SandboxConfigurationComponent } from './pages/sandbox.component';
+import { SandboxConfigurationDetailsComponent } from './component/sandbox-details.component';
 import { ButtonsModule, ModalModule } from 'ngx-bootstrap';
 import { HttpClientModule } from '@angular/common/http';
 import { CommonModule } from '../../shared/common.module';
-import { SandboxService } from './sandbox.service';
-import { NewSandboxConfigurationComponent } from './new-sandbox.component';
+import { SandboxService } from './service/sandbox.service';
+import { NewSandboxConfigurationComponent } from './pages/new-sandbox.component';
 import { TableModule } from 'primeng/table';
 import { MenuModule } from 'primeng/menu';
 import { AccordionModule } from 'primeng/primeng';
-import { DeleteSandboxConfigurationModalComponent } from './modals/delete-sandbox.modal';
-import { ArchiveSandboxConfigurationModalComponent } from './modals/archive-sandbox.modal';
-import { ResetDataModalComponent } from './modals/reset-data.modal';
+import { DeleteSandboxConfigurationModalComponent } from './modal/delete-sandbox.modal';
+import { ArchiveSandboxConfigurationModalComponent } from './modal/archive-sandbox.modal';
+import { ResetDataModalComponent } from './modal/reset-data.modal';
 
 @NgModule({
   declarations: [
